Switched layout from scrollview to frameLayout. What would be better of thinking of adding overlay elements? probably RelativeLayout later on
Images from the slide are loaded with AsyncTask and not on fragment's onCreateView.
It has a length of 5 fragments in memory at once. Good enough I think, and images bitmaps are put into the cache, so memory usage went up (because the parent activities still hold references to their own bitmaps that were problably pushed off the cache...) so total memory is ~ 2 activities ImageView's Bitmap's plus cache

The list of images from the gallery view to the slide view is passed through a singleton. This causes a problem, as sometimes when I go back to the activity, it is Created and somehow the singleton is new and has no reference to an images list whatsoever. Need to check up on that

Slide bitmaps size is calculated from the display size on the activity
GallerySlideActivity.this.getWindowManager().getDefaultDisplay().getSize(preferedSize);
There is probably a better way

Overall it works smoothly on my phone (Galaxy Note 3)