The Exif support for android is pretty incomplete. Comparing to exifftool I don't see nearly as many tags- Android gives me a tag named DateTime while exif speaks aqout a tag named DateTimeOriginal. I guess they are the same, so I will favor that for the date of the images when there is exif. If there is none, the date will be that from the MediaStore (ToDo)
Exif data is loaded at the same time as the image. This results in the data not displaying if it has not been loaded before the user scrolls up to it's image, since the views containing this data are updated only for the current image at "onPageSelected" on the slide activity.
FullScreen now has the SYSTEM_UI_FLAG_IMMERSIVE_STICKY.
This results on the title bar hiding and never showing up. See the docs for an explanation on that.
I don't need that bar anyway since I have the drawer interface.

The image data is presented on a LinearLayout where each element is a RelativeLayout comprised of an ImageView next to a TextView.
This gives opportunity to add icons beside text and a more 'dynamic' data display.